fix: only alter the template for the lit components, nothing else (#138)
[Ultrahtml](https://github.com/natemoo-re/ultrahtml) has some issues with parsing HTML which is specific DSL of Vue. E.g. a scoped slot template like

```html
<template #someslot="{ someProp }">
  <div>{{ someProp }}</div>
</template>
```

is rendered incorrectly by the Ultrahtml renderer.

```html
<template ="}"" someProp="">
<div>{{ someProp }}</div>
</template>
```

Before this PR, we rendered the entire generated AST using Ultrahtml's renderer.

This caused such scoped slots templates to be incorrect after the `autoLitWrapper` plugin was applied. And those templates were not supposed to be altered by the autoLitWrapper at all in the first place.

With this PR, we now surgically alter only the Lit elements configured to be auto-wrapped. Nothing else. And we do this using [magic-string](https://www.npmjs.com/package/magic-string).

This leaves the rest of the template as is. And we do not run into the shortcomings of Ultrahtml.